Fix for remaining RPC messages
